backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4
python-dateutil==2.9.0
pydantic==2.4.2
pydantic-settings==2.0.3 